fix(annotations): move discoverability changes into DocBaseViewer
